Replaced double with decimal
It is recommended to use Decimal over Double for prices (Apple's prices come back as NSDecimalNumber, which is an interchangeable reference semantic version of a Decimal). However, the encoder doesn't like encoding Decimal values so we're now sending them off as strings in the postback.

+++ Sources/Paywall/Models/SW Product/SWProduct.swift
@@ -10,7 +10,7 @@ import StoreKit
 struct SWProduct: Codable {
   var localizedDescription: String
   var localizedTitle: String
-  var price: Double
+  var price: Decimal
   var priceLocale: String
   var productIdentifier: String
   var isDownloadable: Bool
@@ -26,7 +26,7 @@ struct SWProduct: Codable {
   init(product: SKProduct) {
     localizedDescription = product.localizedDescription
     localizedTitle = product.localizedTitle
-    price = product.price.doubleValue
+    price = product.price as Decimal
     priceLocale = product.priceLocale.identifier
     productIdentifier = product.productIdentifier
     isDownloadable = product.isDownloadable
